Invert parameters in EmitterStack
In order to handle stream and non-stream responses Zend Expressive needs
to first try with the SapiStreamEmitter and then the SapiEmitter, since
the latter does not check and simply echo's data.

Since the implementation is a stack, the reading order is inverted, this
we need to insert the stream emitter last, not first.