No longer allow registering multiple applications
In order to simplify the container generation process, let's stop allowing the
registration of multiple apps in a container and stick to a single one.

A side effect of these changes is that we can statically create the XML files
for the services created in the compiler passes. We looked into it but we
couldn't do it in a time-efficient manner.